+++ lemoncheesecake/__init__.py
@@ -1,3 +1,3 @@
-VERSION_INFO = (1, 6, 0)
+VERSION_INFO = (1, 7, 0)
 
 __version__ = '.'.join(map(str, VERSION_INFO))