+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 - **API**: parametrized test improvements:
   - parameters can now also be passed in a CSV-like mode
-  - support strings with placeholders as a naming scheme
+  - add support for strings with placeholders as a naming scheme
 - **CLI**: the `LCC_PROJECT_FILE` environment variable has been deprecated/renamed in favor of `LCC_PROJECT`
 - **CLI**: the `LCC_PROJECT` environment variable can also be used with a directory that holds either a `project.py` file or a `suites` sub-directory
 - **CLI**: `lcc run` has a new `--project/-p` argument that takes a project path
+- **API**: add matcher `has_all_items()`
 - **API**: Fix `has_entry` matcher exception when an item cannot be found by index like in `has_entry(("foo", 0))` with the
   actual data being `{"foo": []}`; also fixes the same use case with `{check,require,assert}_that_in` functions
 - **API**: Fix `starts_with/ends_with/contains_string` matchers exception when the actual value is not a string